+++ Code/CTOMDataPersistence-PG/src/gov/nih/nci/ctom/ctlab/domain/Specimen.java
@@ -97,7 +97,7 @@ public class Specimen
 	private String commentsFromInvestigator = null;
 	private Long sampleTypeCdId;
 	private Long specimenCollectionId;
-	private int securityKey;
+	private Long securityKey;
 	private String source;
 	private Date srcExtractDt;
 	private Date ctomInsertDt;
@@ -364,7 +364,7 @@ public class Specimen
 	/**
 	 * @return the securityKey
 	 */
-	public int getSecurityKey()
+	public Long getSecurityKey()
 	{
 		return securityKey;
 	}
@@ -373,7 +373,7 @@ public class Specimen
 	 * @param securityKey
 	 *            the securityKey to set
 	 */
-	public void setSecurityKey(int securityKey)
+	public void setSecurityKey(Long securityKey)
 	{
 		this.securityKey = securityKey;
 	}
